move modes to lib
