+++ src/components/mode-tools/mode-tools.jsx
@@ -13,7 +13,7 @@ import Input from '../forms/input.jsx';
 import InputGroup from '../input-group/input-group.jsx';
 import LabeledIconButton from '../labeled-icon-button/labeled-icon-button.jsx';
 // import LabeledIconButton from '../labeled-icon-button/labeled-icon-button.jsx';
-import Modes from '../../modes/modes';
+import Modes from '../../lib/modes';
 import styles from './mode-tools.css';
 
 import copyIcon from './icons/copy.svg';
