+++ src/lib/modes.js
